% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{BigSyn}
\alias{BigSyn}
\title{BigSyn: Some non confidential functions developped for the MLDSCR synthetic data project}
\description{
The BigSyn package allows to synthesize big hierarchical databases by opposition to just a single small "rectangular" table.
}
\section{BigSyn functions}{

The BigSyn functions are
}

