% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilefits.R
\name{compilefits}
\alias{compilefits}
\title{compilefits}
\usage{
compilefits(
  Sparameters,
  fitmodelsavepath,
  pdfpath = fitmodelsavepath,
  .progress = "text"
)
}
\arguments{
\item{pdfpath}{where to save the pdfs}

\item{Sparameters:}{a list, that has the same structure than the outputs of}

\item{fitmodelsavepath:}{a file path where to store the pdf of the plot}

\item{.progress:}{a string, name of the progress bar to use, see plyr::create_progress_bar}
}
\description{
For each element of save parameters, look at the tree and produces the corresponding pdf.  It also removes all the information that is stored in the ouptut of parykit::Ctree, e.g. the data. It only keeps the tree and the rules to get it.
}
\details{
depends on plyr
}
\examples{
y=iris$Species;x=iris[,-5]
partykitctree <- partykit::ctree(y ~ ., data=cbind(y=y,x))
}
