% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilefits.R
\name{compilesamplereports}
\alias{compilesamplereports}
\title{compilesamplereports}
\usage{
compilesamplereports(Sparameters, samplereportssavepath)
}
\arguments{
\item{Sparameters:}{a list, that has the same structure than the outputs of}

\item{samplereportssavepath:}{a file path where to store the sample reports}
}
\description{
Sample reports are the output of the function ReportonSample
}
\details{
depends on plyr
}
\examples{
y=iris$Species;x=iris[,-5]
partykitctree <- partykit::ctree(y ~ ., data=cbind(y=y,x))
}
\seealso{
ReportonSample
}
