% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daniRules.R
\name{daniRules}
\alias{daniRules}
\title{daniRules}
\usage{
daniRules(x, i = NULL, ...)
}
\arguments{
\item{x}{same format than output of partykit::ctree}

\item{i}{(default: NULL)}
}
\value{
for each terminal node of the tree, give the definition. For example: node 11 corresponds to 'x>1 and y in ("a","b")'
}
\description{
the rules for the tree. Add an extra rule. For example, if the left branch rule is X in (a,b,c) and the right branch rule is X in (e,f),
the right branch rule is modified by X not in (a,b,c). The extreme right branch rule is replaced by the negation of any of the other branches, for each node.
}
\details{
Adapted from partykit:::.list.rules.party
}
\examples{
x=partykit::ctree(Petal.Width~.,iris)
daniRules(x)   
}
