% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_synthesised_data.R
\name{donors.receptors.check}
\alias{donors.receptors.check}
\title{For each leave of each tree, counts the number of donors in the gold set  
vs the number of receptors in the synthetic dataset.}
\usage{
donors.receptors.check(Rules, gold.data, syn.data)
}
\arguments{
\item{Rules}{a named list (names correspond to the gold.data and syn.data variable names) of lists of logical expressions (for each variable, the logical expression corresponds to a terminal leaf in a tree).}

\item{gold.data}{a data frame}

\item{syn.data}{a data frame containing the same variables than gold.data}
}
\description{
For each leave of each tree, counts the number of donors in the gold set  
vs the number of receptors in the synthetic dataset.
}
\examples{
data(TtableA,package="BigSyn")
STtableA1=BigSyn::SDPSyn2(TtableA,asis=c("id1","id2"))
data(STtableA1,package="BigSyn")
data(package="BigSyn")
donors.receptors.check()
}
