% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{drop_last}
\alias{drop_last}
\title{Drop last marginpos (cuts everything after last after "_")}
\usage{
drop_last(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Drop last marginpos (cuts everything after last after "_")
}
\details{
if x is "aa.x_a_1_f_1" returns "a_1_f_1"
}
\examples{
drop_last("AA.char1_La_Ld_Lrn1")
}
