% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRpackagedata_to_csvlibrary.R
\name{exportRpackagedata_to_csvlibrary}
\alias{exportRpackagedata_to_csvlibrary}
\title{export all tables from an installed R data package to sas7bdat data files in a given library}
\usage{
exportRpackagedata_to_csvlibrary(package, path_to_export_to = getwd(),
  tables = NULL, zip = TRUE)
}
\arguments{
\item{package}{package to be transformed}

\item{path_to_export_to}{path to export the data}

\item{tables}{tables to extract, if NULL, all tables.}

\item{zip}{(should the files be zipped (requires utils package))}
}
\value{
nothing
}
\description{
export all tables from an installed R data package to sas7bdat data files in a given library
}
\examples{
exportRpackagedata_to_csvlibrary("datasets")
}
