% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{fitmodel.fn}
\alias{fitmodel.fn}
\title{Fit a model with a specific function}
\usage{
fitmodel.fn(method, x, y, treeplotsavepath = NULL, ...)
}
\arguments{
\item{method}{a string. currently only method="ctree" or "rf" (random forest).}

\item{x}{a predictors, a dataframe.}

\item{y}{variable to predict, a vector}

\item{treeplotsavepath}{a}

\item{...}{synthetic parameters to pass to the right fit model function. the fit model function name is the concatenation of "fit.model" and method}
}
\value{
a sublist of synparameters, which names are possible arguments of synthpop::syn.ctree if method="ctree".
}
\description{
Fit a model with a specific function
}
\examples{
fitmodel.fn(method="ctree",x=iris[,-5],y=iris$Species,nbuckets=30,tutu="not a good argument")
}
