% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{fitmodel.rf}
\alias{fitmodel.rf}
\title{Function to fit a ctree model.}
\usage{
fitmodel.rf(x, y, treeplotsavepath = NULL, ...)
}
\arguments{
\item{x}{a dataframe of predictors}

\item{y}{a vector :dependent variable}

\item{treeplotsavepath:}{a path to save the graph as a pdf. if NULL, no pdf is saved}
}
\value{
a named list of 4 elements:
"Rules" a data.frame with two variables: terminalnode (a integer vector) and condition  a string that gives the rule for each terminal node.
 "y" the values of the predictor
 "terminalnodes" a vector: the terminal nodes for each element of $y$.
 "shortlist" a character string giving the names of the variables in x that were used for the classification
}
\description{
Function to fit a ctree model.
}
\examples{
fitmodel.ctree(x=iris[,-5],y=iris$Species)
}
