% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{fitthemodel}
\alias{fitthemodel}
\title{Function to fit the model.}
\usage{
fitthemodel(
  Sparameters_i,
  fitmodelsavepath,
  TtableANAto0,
  redocomputationsevenifexists = FALSE,
  treeplotsavefolder = NULL
)
}
\arguments{
\item{Sparameters_i}{an element of the list output from Sparameters.default.f}

\item{fitmodelsavepath}{a folder path. Results will either be read from or stored in this folder. If the file exists, by default it is not replaced.}

\item{TtableANAto0}{a table  containing the predictors without NAs as well as the outcome}
}
\value{
a list.
}
\description{
Function to fit the model.
}
\examples{

data(TtableA,package="BigSyn")
Sparameters<-Sparameters.default.f(ref.table=TtableA)
Sparameters_i<-Sparameters[[53]]; fitmodelsavepath=NULL; TtableANAto0<-NAto0(TtableA);redocomputationsevenifexists=FALSE
treeplotsavefolder=tempdir()
fitthemodel(Sparameters_i,fitmodelsavepath = NULL,TtableANAto0 = TtableANAto0,treeplotsavefolder=tempdir())
Sparameters_i<-Sparameters[["AA.present_La_La_Lrn1"]]; 
treeplotsavefolder=tempdir()
fitthemodel(Sparameters_i,NULL,TtableANAto0,treeplotsavefolder=tempdir())
}
