% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTransposingFunctions.R
\name{GeneralReversetransposefunction}
\alias{GeneralReversetransposefunction}
\title{General Reverse Transpose function}
\usage{
GeneralReversetransposefunction(TtableA, key)
}
\arguments{
\item{key}{A list of variables (columns of the transposed table)}

\item{table}{A dataframe}
}
\value{
A list: first element of the list is a dataframe, the transposed version of the orioginal table. Second element is a key to allow back transposition
}
\description{
General Reverse Transpose function
}
\examples{
data(tableA);data(TtableA);data(XKA);key<-XKA$key
RtableA=GeneralReversetransposefunction(TtableA,key)
ordertableA <-do.call(order,tableA[c(id1,id2)])
orderRtableA<-do.call(order,RtableA[c(id1,id2)])
identical(nrow(tableA),nrow(RtableA))
identical(lapply(tableA,class),lapply(RtableA,class))
identical(tableA[ordertableA,],RtableA[orderRtableA,])
identical(names(tableA),names(RtableA))
all (lapply(names(tableA),function(x){identical(tableA[ordertableA,x],RtableA[orderRtableA,x])}))
}
