% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTransposingFunctions.R
\name{GeneralReversetransposefunctiondecoupe}
\alias{GeneralReversetransposefunctiondecoupe}
\title{General Reverse Transpose function with split}
\usage{
GeneralReversetransposefunctiondecoupe(.data, key, nrowmax = 10000)
}
\arguments{
\item{key}{A list of variables (columns of the transposed table)}

\item{table}{A dataframe}
}
\value{
A list: first element of the list is a dataframe, the transposed version of the orioginal table. Second element is a key to allow back transposition
}
\description{
General Reverse Transpose function with split
}
\examples{
data(tableA);data(TtableA);data(XKA);key<-XKA$key
RtableA=GeneralReversetransposefunctiondecoupe(TtableA,key,10)
}
