% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTransposingFunctions.R
\name{Generaltransposefunction}
\alias{Generaltransposefunction}
\title{General Transpose function}
\usage{
Generaltransposefunction(
  tableA,
  id1,
  id2,
  origin = deparse(substitute(tableA))
)
}
\arguments{
\item{id1}{A list of variables (rows)}

\item{id2}{A list of variables (columns of the transposed table), id2 can contain as a last element the strint "rn", if the variable rn is an index for the cells formed by the variables listed first in id2}

\item{table}{A dataframe}
}
\value{
A list: first element of the list is a dataframe, the transposed version of the orioginal table. Second element is a key to allow back transposition
}
\description{
General Transpose function
}
\examples{
tableA<-sampledata(TRUE)
id1=c("id1a","id1b")
id2=c("id2a","id2b")
TtableA<-Generaltransposefunction(tableA,id1,id2)
}
