% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTransposingFunctions.R
\name{Generaltransposefunctionsimple}
\alias{Generaltransposefunctionsimple}
\title{Simple General Transpose function}
\usage{
Generaltransposefunctionsimple(tableA, id1, id2)
}
\arguments{
\item{tableA}{A dataframe}

\item{id1}{A list of variables (rows)}

\item{id2}{A list of variables (columns of the transposed table)}
}
\value{
A data frame
}
\description{
Simple General Transpose function
}
\examples{
tableA<-sampledata(TRUE)
id1=c("id1a","id1b")
id2=c("id2a","id2b")
TtableA<-Generaltransposefunctionsimple(tableA,id1,id2)
}
