% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_cell}
\alias{get_cell}
\title{get cell without the row number}
\usage{
get_cell(x, iscellrn = FALSE, iscell = FALSE)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
get cell without the row number
}
\details{
if x is "aa.xoijj_a_1_f_1_" returns "a_1_f"
}
\examples{
get_cell("aa.x_1_2_3_4")#default
get_cell("1_2_3",TRUE)
get_cell("1_2_3",FALSE,TRUE)
unique(Tsampledata(TRUE)$variables))
unique(get_cell(Tsampledata(FALSE)$variables))
}
