% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_cellrn}
\alias{get_cellrn}
\title{Get cell and row number}
\usage{
get_cellrn(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Get cell and row number
}
\details{
if x is "aa.x_a_1_f_1" returns "a_1_f_1"
}
\examples{
get_cellrn("AA.char1_La_Ld_Lrn1")
data(TtableA);
unique(get_cellrn(names(TtableA)))
#Second example: no transposing variables
data(TtableB);data(XKB)
unique(get_cellrn(names(XKB)))
}
