% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_cellXXgroup}
\alias{get_cellXXgroup}
\title{Get cell group}
\usage{
get_cellXXgroup(x, marginpos, iscellXX = TRUE)
}
\arguments{
\item{x}{a vector of character strings}

\item{marginpos}{a vector of integer}
}
\value{
a vector of character strings
}
\description{
Get cell group
}
\details{
if x is "a_1_f_2_aa.xoijj",marginpos=2 returns "1"
if x is "a_1_f_2_aa.xoijj",marginpos=-2 returns "a_f_2"
if x is "a_1_f_2_aa.xoijj",marginpos=c(1:2) returns "a_1"
}
\examples{
get_cellXXgroup(c("aa.x_1_2_3_4","bb.x_1_2_3_4"),2,iscellXX=FALSE)
get_cellXXgroup(c("1_2_3_4","1_2_3_4"),2:3,iscellXX=TRUE)
variables<-Tsampledata(TRUE)$variables
unique(get_cellXXgroup(variables,2,iscellXX=FALSE))
unique(get_cellXXgroup(variables,-2,iscellXX=FALSE))
get_cellXXgroup(variables[50],2,iscellXX=FALSE)
get_cellXXgroup(variables[50],-2,iscellXX=FALSE)

#Second example: no transposing variables
TK<-Tsampledata(FALSE)
unique(get_cellXXgroup(TK$variable,1,iscell=FALSE))
}
