% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_cellXXmarginscount}
\alias{get_cellXXmarginscount}
\title{Get the number of margins for a cell}
\usage{
get_cellXXmarginscount(x, iscellXX = FALSE)
}
\arguments{
\item{x}{a vector of character strings}

\item{iscell}{a boolean indicating if x is a variable name or a cell name.}
}
\value{
a vector of  integers.
}
\description{
Get the number of margins for a cell
}
\details{
if x is "aa.xoijj_a_1_f_1", cell=FALSE returns 4"
if x is "a_1_f_1", cell=TRUE returns 4"
}
\examples{
get_cellXXmarginscount("1_2_3_4",iscellXX=TRUE)
get_cellXXmarginscount("aa.x_1_2_3_4",iscellXX=FALSE)
data(TtableA)
unique(get_cellXXmarginscount(names(TtableA),iscellXX=FALSE))
#Second example: no transposing variables
TK<-Tsampledata(FALSE)
unique(get_cellXXmarginscount(TK$variables))
}
