% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_cellXXsplit}
\alias{get_cellXXsplit}
\title{split a cell}
\usage{
get_cellXXsplit(x, marginpos = NULL, iscellXX = FALSE)
}
\arguments{
\item{x}{a vector of character strings}

\item{iscell}{x a boolean indicating if x is a cell}
}
\value{
a vector of character strings
}
\description{
split a cell
}
\details{
if x is "aa.xoijj_a_1_f_1" returns c("a","1","f","1")
}
\examples{
get_cellXXsplit("aa.x_1_2_3_4",iscellXX=FALSE)
get_cellXXsplit("1_2_3_4",iscellXX=TRUE)
get_cellXXsplit("1_2_3_4",2:3,iscellXX=TRUE)
get_cellXXsplit("1_2_3_4",-(2:3),iscellXX=TRUE)
variables<-Tsampledata(TRUE)$variables
unique(get_cellXXsplit(variables,iscell=FALSE))
get_cellXXsplit(variables[50],iscell=FALSE)
get_cellXXsplit(variables[50],-(2:3),iscell=FALSE)
unique(get_cellXXsplit(variables,2,iscell=FALSE))
#Second example: no transposing variables
TK<-Tsampledata(FALSE)
unique(get_cellXXsplit(TK$variables,iscell=FALSE))
}
