% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_missingind}
\alias{get_missingind}
\title{Get missing indicator for a cell or variable}
\usage{
get_missingind(x, variables)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Get missing indicator for a cell or variable
}
\details{
if x is "a_1_f_1_aa.xoijj" returns c("a","1","f","1")
}
\examples{
variables<-Tsampledata(TRUE)$variables
unlist(unique(get_missingind(variables,variables)))
variables<-Tsampledata(FALSE)$variables
unlist(unique(get_missingind(variables,variables)))
}
