% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_natural.predictors}
\alias{get_natural.predictors}
\title{get variable predecessors at margin}
\usage{
get_natural.predictors(x, variables = x, predictors = NULL)
}
\arguments{
\item{x}{a vector of character strings}

\item{variables}{a vector of character strings}

\item{cells}{a vector of character strings containing the potential predecessors}

\item{marginpos}{a vector of integers}

\item{x}{a vector of character strings}
}
\value{
a vector of character strings

a vector of character strings
}
\description{
get variable predecessors at margin
}
\details{
if x is "a_1_f_1_aa.xoijj" returns c("a","1","f","1")

if x is "a_1_f_1_aa.xoijj" returns c("a","1","f","1")
}
\examples{
get_XXpredecessoratmargin(cellXXs="aa.x_1_2_3_4", refcellXXs=c("bb.x_1_2_2_4","aa.x_1_2_2_4","aa.x_1_1_3_4"),2,iscellXX=FALSE)
get_XXpredecessoratmargin(cellXXs=c("1_2_2_4","1_2_2_4","1_1_3_4","1_1_3_3"),iscellXX=FALSE)
data(XKA)
cells<-unique(get_cellrn(XKA$variables))
get_XXpredecessoratmargin(cells,marginpos=1,iscellXX=TRUE)
get_XXpredecessoratmargin(cells[10],cells,1,iscellXX=TRUE)
Get natural predictors

TK<-TtableA
get_natural.predictors(x=sample(names(TtableA),5),variables=names(TtableA))
}
