% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_origin}
\alias{get_origin}
\title{Get origin table}
\usage{
get_origin(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Get origin table
}
\details{
if x is "aa.xoijj_a_1_f_1_" returns c("aa")
}
\examples{
get_origin("tableA.cont1_1_Lrn1")
variables<-Tsampledata(TRUE)$variables
unlist(unique(get_origin(variables,variables)))
variables<-Tsampledata(FALSE)$variables
unlist(unique(get_origin(variables,variables)))
}
