% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_parent}
\alias{get_parent}
\title{get parent if any}
\usage{
get_parent(variables, variable_ref)
}
\arguments{
\item{variables}{a character strings}

\item{variable_ref}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
get parent if any
}
\details{
if x is "aa.xoijj_a_1_f_1_" returns "a_1_f"
}
\examples{
get_parent("aa.x_1_2_3_4","aa.x_1_2_3")#default
}
