% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_presentind}
\alias{get_presentind}
\title{get the present indicator for a cell}
\usage{
get_presentind(variables, refvariables = variables,
  rns = unlist(unique(get_cellrn(refvariables))))
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
get the present indicator for a cell
}
\details{
if x is "a_1_f_1_aa.xoijj" returns c("a","1","f","1")
}
\examples{
get_presentind("AA.x_1_2_3_4","AA.present_1_2_3_4")
get_presentind("AA.present_1_2_3_4",c("AA.present_1_2_3_3","AA.present_1_2_3_4"))
variables<-Tsampledata(TRUE)$variables
variable<-"AA.present_La_La_Lrn1"
get_presentind(variable,variables)
unlist(unique(get_presentind(variables)))
variables<-Tsampledata(FALSE)$variables
unlist(unique(get_presentind(variables,variables)))
}
