% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_var}
\alias{get_var}
\title{Get variable name}
\usage{
get_var(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Get variable name
}
\details{
if x is "aa.xoijj_a_1_f_1" returns "aa.xoijj"
}
\examples{
get_var("aa.x_1_2_3_4")
data(TtableA)
unique(get_var(names(TtableA)))
#Second example: no transposing variables
TK<-Tsampledata(FALSE)
unique(get_var(TK$variables))
}
