% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{getnodesfromrules}
\alias{getnodesfromrules}
\title{Function to get terminal node from a set of partitioning rules and new predictors}
\usage{
getnodesfromrules(x, Rules)
}
\arguments{
\item{x}{a dataframe of predictors}

\item{Rules}{a data frame containing 2 character variables: "terminalnode" and "condition"}
}
\value{
a vector of lenth the number of rows of x indicating the terminal nodes.
}
\description{
Function to get terminal node from a set of partitioning rules and new predictors
}
