% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpredictorsfromtree.R
\name{getpredictorsfromtree}
\alias{getpredictorsfromtree}
\title{getpredictorsfromtree}
\usage{
getpredictorsfromtree(tree, predictors)
}
\arguments{
\item{tree}{same format than output of partykit::ctree}

\item{predictors}{list of variables}
}
\value{

}
\description{
for each terminal node of the tree, give the elements of "predictors" who appear in the node definition
}
\examples{
tree<-partykit::ctree(Petal.Width~.,iris)
getpredictorsfromtree(tree,names(iris))
}
\seealso{
daniRules
}
