% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{good.fit.parameters}
\alias{good.fit.parameters}
\title{Select only arguments for a specific fitting function}
\usage{
good.fit.parameters(method, synparameters)
}
\arguments{
\item{method}{a string. currently only method="ctree".}

\item{synparameters}{a named list.}
}
\value{
a sublist of synparameters, which names are possible arguments of partykit::ctree_control if method="ctree".
}
\description{
Select only arguments for a specific fitting function
}
\details{
Currently only works with method="ctree"
Only selects the arguments that match the function partykit::ctree_control
}
\examples{
good.fit.parameters(method="ctree",list(teststat=30,tutu=3))
}
