% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{good.sample.parameters}
\alias{good.sample.parameters}
\title{Select only arguments for sampling.}
\usage{
good.sample.parameters(method, synparameters)
}
\arguments{
\item{method}{a string. currently only method="ctree".}

\item{synparameters}{a named list.}
}
\description{
Select only arguments for sampling.
}
\details{
In prevision of future developments. returns NULL for the moment
}
\examples{
good.sample.parameters(method="ctree",list(teststat=30,tutu=3))
}
