% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{good.syn.parameters}
\alias{good.syn.parameters}
\title{Select only arguments for a specific fitting function
This function is not used anymore}
\usage{
good.syn.parameters(method, synparameters)
}
\arguments{
\item{method}{a string. currently only method="ctree".}

\item{synparameters}{a named list.}
}
\value{
a sublist of synparameters, which names are possible arguments of synthpop::syn.ctree if method="ctree".
}
\description{
Select only arguments for a specific fitting function
This function is not used anymore
}
\details{
Currently only works with method="ctree"
Only selects the arguments that match the function synthpop::syn.ctree
}
\examples{
good.syn.parameters(method="ctree",list(y=c(1:2),smoothing=TRUE,tutu="not in synthpop::syn.ctree arguments"))
}
