% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{NAto0}
\alias{NAto0}
\title{General Default ordering of variables for synthetisation based on name of the variable.}
\usage{
NAto0(tableA)
}
\arguments{
\item{tableA}{a dataframe}
}
\value{
a dataframe
}
\description{
General Default ordering of variables for synthetisation based on name of the variable.
}
\examples{
toto<-cars
toto$speed[sample(nrow(cars),3)]<-NA
NAto0(toto)
}
