% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{onlygoodargs}
\alias{onlygoodargs}
\title{Generic function: remove all the elements of a named list which names are not arguments of a specific function.}
\usage{
onlygoodargs(fun, L)
}
\arguments{
\item{fun}{a function}

\item{L}{a list}
}
\value{
a list.
}
\description{
Generic function: remove all the elements of a named list which names are not arguments of a specific function.
}
\details{
remove all the elements of a list which names are not arguments of a specific function.
}
\examples{
onlygoodargs(lm,list(data=cars,formula=speed~dist,tutu="not an arg from lm"))
}
