% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{posixcttonumeric}
\alias{posixcttonumeric}
\title{Converts all posixct variables of a dataframe into a numeric variable}
\usage{
posixcttonumeric(tableA)
}
\arguments{
\item{tableA}{a dataframe}
}
\value{
a list.
}
\description{
Converts all posixct variables of a dataframe into a numeric variable
}
\examples{
toto<-cars
toto$now<-Sys.time()
posixcttonumeric(toto)
}
