% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{predictor.matrix.default}
\alias{predictor.matrix.default}
\title{Define a default predictor matrix}
\usage{
predictor.matrix.default(variables)
}
\arguments{
\item{variables}{a vector of character strings}
}
\value{
a matrix
}
\description{
Define a default predictor matrix
}
\details{
Returns the lower diagonal matrix with ones.
}
\examples{
variables<-Tsampledata(TRUE)$variables
predictor.matrix.default(TK$variables)
}
