% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{predictor.matrix.rate}
\alias{predictor.matrix.rate}
\title{predictor.matrix.rate}
\usage{
predictor.matrix.rate(
  variables,
  nopredictor = character(0),
  allpredictor = character(0),
  marginposs = integer(0)
)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
predictor.matrix.rate
}
\details{
if x is "aa.xoijj_a_1_f_1_" returns c("a","1","f","1")
}
