% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{preparepredictorsforctreefit}
\alias{preparepredictorsforctreefit}
\title{Prepare predictors for ctree fit}
\usage{
preparepredictorsforctreefit(x, keep = NULL)
}
\arguments{
\item{x}{a predictors, a dataframe.}

\item{method}{a string. currently only method="ctree".}

\item{y}{variable to predict, a vector}

\item{...}{synthetic parameters to pass to the right fit model function. the fit model function name is the concatenation of "fit.model" and method}
}
\value{
a sublist of synparameters, which names are possible arguments of synthpop::syn.ctree if method="ctree".
}
\description{
Prepare predictors for ctree fit
}
\examples{
preparepredictorsforctreefit(x,
                             keep=NULL)
}
