% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleT_f.R
\name{resampleT_f}
\alias{resampleT_f}
\title{resample  when synthetisation created incoherent high and low level aggregates}
\usage{
resampleT_f(.data, variables, verbose = FALSE)
}
\arguments{
\item{.data}{data frame to "reduce"}

\item{variables}{list of variable names roots}

\item{verbose}{(default FALSE) if verbose, the formulae to compute the new variables is printed.}
}
\description{
resample  when synthetisation created incoherent high and low level aggregates
}
