% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{sample.ctree}
\alias{sample.ctree}
\title{Function to sample from a ctree fitted model}
\usage{
sample.ctree(xp, fit.model, smoothing = "none", ...)
}
\arguments{
\item{y}{a vector of values to pull from}

\item{terminalnodes}{a vector of terminal nodes}

\item{newterminalnodes:}{a path to save the graph}
}
\value{
a vector of the same size than terminalnodes, obtained by sampling betweenn the values of y such for the same terminal node.
}
\description{
Function to sample from a ctree fitted model
}
\examples{
y<-iris$Species;x<-xp<-iris[,-5];fit.model<-fitmodel.ctree(x,y);sample.ctree(x,fit.model)
}
