% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{sample.fn}
\alias{sample.fn}
\title{Sample a model with a specific function}
\usage{
sample.fn(method, xp, fit.model, smoothing, ...)
}
\arguments{
\item{method}{a string. currently only method="ctree".}

\item{...}{synthetic parameters to pass to the right fit model function. the fit model function name is the concatenation of "fit.model" and method}

\item{x}{a predictors, a dataframe.}

\item{y}{variable to predict, a vector}
}
\value{
a sublist of synparameters, which names are possible arguments of synthpop::syn.ctree if method="ctree".
}
\description{
Sample a model with a specific function
}
\examples{
sample.fn(method="ctree",
          xp=iris[,-5],
          fit.model=fitmodel.fn(method="ctree",x=iris[,-5],y=iris$Species,nbuckets=30),
          smoothing=FALSE)
}
