% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\name{sampledata}
\alias{sampledata}
\title{Sample data for transposition}
\usage{
sampledata(transposingvariables = TRUE)
}
\arguments{
\item{transposingvariables}{a boolean. If TRUE, transposing  id variables are created.}
}
\value{
a data frame with id variables, numeric, factor and character variables.
}
\description{
Sample data for transposition
}
