% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDPSyn.R
\name{samplefrompool}
\alias{samplefrompool}
\title{Function to sample from a set of partitioning rules}
\usage{
samplefrompool(y, terminalnodes, newterminalnodes)
}
\arguments{
\item{y}{a vector of values to pull from}

\item{terminalnodes}{a vector of terminal nodes}

\item{newterminalnodes:}{a path to save the graph}
}
\value{
a vector of the same size than terminalnodes, obtained by sampling betweenn the values of y such for the same terminal node.
}
\description{
Function to sample from a set of partitioning rules
}
\examples{
y=iris$Species;x=iris[,-5];fit.mod<-fitmodel.ctree(x,y);terminalnodes<-getnodesfromrules(x,fit.mod$Rules);
newterminalnodes<-sample(unique(terminalnodes),10,replace=TRUE);
samplefrompool(y,terminalnodes,newterminalnodes)
y<-y[terminalnodes!=7]
terminalnodes<-terminalnodes[terminalnodes!=7]
samplefrompool(y,terminalnodes,newterminalnodes)
}
