% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{SDPSYN}
\alias{SDPSYN}
\title{General SDP function.}
\usage{
SDPSYN(TtableA, asis = NULL, notpredictor = asis, replicate = 1,
  Sparameters = Sparameters.default.f(TtableA, asis, notpredictor),
  STtableA = plyr::rdply(replicate, TtableA[asis]))
}
\description{
General SDP function.
}
\examples{

TK<-Tsampledata(TRUE)
TtableA<-TK$TtableA;key=TK$key;
STtableA=as.data.frame(matrix(NA,nrow(TtableA),0))
Sparameters=Sparameters.default.f(ref.table=TtableA)
Sparameters_i<-Sparameters[[54]]
Split=Sparameters_i$split[[1]]
Split=Sparameters_i$split[[2]]
library(synthpop)
TtableA[sapply(TtableA,is.character)]<-lapply(TtableA[sapply(TtableA,is.character)],as.factor)
STtableA<-SDPSYN(TtableA)
StableA<-BigSyn::GeneralReversetransposefunction(STtableA,key) 
}
