% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{sorttablewithingroup}
\alias{sorttablewithingroup}
\title{sort table within group without changing the group position}
\usage{
sorttablewithingroup(.data, groupvar, sortvar, decreasing = FALSE)
}
\arguments{
\item{.data}{a dataframe}

\item{groupvar}{a vector of character strings that are names of variables from .data}

\item{sortvar}{a vector  of character strings that are names of variables from .data}

\item{decreasing}{a boolean if TRUE, decreasing order is used.}
}
\value{
a list.
}
\description{
sort table within group without changing the group position
}
\details{
Groups are defined by unique values of .data[groupvar]. Within each group, data is sorted according to sortvar.
}
\examples{
set.seed(1)
N=10
.data=data.frame(
 .group=sample(letters[1:2],N,replace=TRUE),
 y=runif(N),
 origorder=1:N)
 groupvar=".group"
 sortvar="y"
.data2=plyr::ddply(.data,".group",function(d){d$intraorder=order(d[[sortvar]]);d$neworder=d$origorder[order(order(d[[sortvar]]))];d})
.data4=.data2[.data2[[groupvar]]=="a",];.data4[.data4$intraorder,]
cbind(.data,"|",.data2)
cbind(.data,"|",.data2[order(.data2$origorder),])
cbind(.data[order(order(.data2$origorder)),],"|",.data2)
.data3=cbind(.data,"|",.data2[order(.data2$neworder),]);.data3[.data[[groupvar]]=="a",];.data3
.data3=cbind(.data,"|",.data[order(order(.data2$origorder)),][order(.data2$neworder),]);.data3[.data[[groupvar]]=="a",]
.data3=cbind(.data,"|",.data[order(order(.data2$origorder))[order(.data2$neworder)],]);.data3[.data[[groupvar]]=="a",]
cbind(.data,I="|",sorttablewithingroup(.data,groupvar,sortvar),I="|",sorttablewithingroup(.data,groupvar,sortvar,decreasing=TRUE),I="|",sorttablewithingroup(.data,NULL,groupvar,decreasing=TRUE))
}
