% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{Sparameters.default.f}
\alias{Sparameters.default.f}
\title{Default synthetisation parameters based on variable names}
\usage{
Sparameters.default.f(ref.table, asis = NULL, notpredictor = NULL,
  variables = Sparameters.variables.reorder.default(names(ref.table)),
  predictors.matrix = predictor.matrix.default(variables)[!is.element(variables,
  asis), !is.element(variables, notpredictor)], moresplits = NULL,
  preferredmethod = "rf")
}
\arguments{
\item{a}{vector of character strings}
}
\description{
Default synthetisation parameters based on variable names
}
\examples{
data(TtableA)
ref.table<-TtableA
Sparameters.default.f(ref.table=TtableA)
}
