% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{Sparameters.variables.reorder.default}
\alias{Sparameters.variables.reorder.default}
\title{General Default ordering of variables for synthetisation based on name of the variable.}
\usage{
Sparameters.variables.reorder.default(
  variables,
  orderwithinorigin = NULL,
  id = NULL,
  extrasort = NULL
)
}
\arguments{
\item{variables}{vector of character strings, indicating names of variables}

\item{orderwithinorigin}{a list, see example}
}
\value{
a list.
}
\description{
General Default ordering of variables for synthetisation based on name of the variable.
}
\examples{
TK<-Tsampledata(TRUE)$TtableA
Sparameters.variables.reorder.default(names(TK$TtableA))
#Second example: no transposing variables
TtableA<-Tsampledata(TRUE)$TtableA
orderwithinorigin=c("AA.factor1","AA.factor2")
variables<-names(TtableA)
Sparameters.variables.reorder.default(variables,orderwithinorigin)
}
