% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilefits.R
\name{treedepth}
\alias{treedepth}
\title{Compute depth of a "party" tree}
\usage{
treedepth(x)
}
\arguments{
\item{x}{a tree}
}
\description{
this function takes the output of the partykit::ctree function and prints the tree into a pdf file, located in the specified folder. It computes the depth and width and tries to create a pdf with the right dimensions.
}
\details{
recursive function
}
\examples{
y=iris$Species;x=iris[,-5]
partyctree <- party::ctree(y ~ ., data=cbind(y=y,x))
treedepth(partyctree@tree)
partyctree <- party::ctree(y ~ ., data=cbind(y=y,x))
treedepth(partykit::ctree(y ~ ., data=cbind(y=y,x)))
}
