% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilefits.R
\name{treetopdf}
\alias{treetopdf}
\title{Ctree to pdf graph}
\usage{
treetopdf(partykitctree, savepath)
}
\arguments{
\item{partykitctree:}{an output of partykit::ctree}

\item{savepath:}{a file path where to store the pdf of the plot}
}
\description{
this function takes the output of the partykit::ctree function and prints the tree into a pdf file, located in the specified folder. It computes the depth and width and tries to create a pdf with the right dimensions.
}
\examples{
y=iris$Species;x=iris[,-5]
partykitctree <- partykit::ctree(y ~ ., data=cbind(y=y,x))
treetopdf(partykitctree,"./x.pdf")
}
