% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\name{TTsampledata}
\alias{TTsampledata}
\title{Transposed sample data.}
\usage{
TTsampledata(transposingvariables = TRUE)
}
\arguments{
\item{transposingvariables}{a boolean. If TRUE, stransposing  id variables are created.}
}
\value{
a data frame with id variables, numeric, factor and character variables.
}
\description{
Transposed sample data.
}
\details{
Tsampledata(x) is Generaltransposefunction(Tsampledata(x))
}
