% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{drop_last}
\alias{drop_last}
\title{Drop last margin position (Trims all strings of a vector of strings after the last "_")}
\usage{
drop_last(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Drop last margin position (Trims all strings of a vector of strings after the last "_")
}
\details{
if x is "AA.char1_La_Ld_Lrn1" returns "AA.char1_La_Ld", if x contains no "_", returns empty string
}
\examples{
drop_last("AA.char1_La_Ld_Lrn1")
drop_last("iojoij")
drop_last("aa.iojoij")
}
