% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{get_natural.predictors}
\alias{get_natural.predictors}
\title{Get natural predictors}
\usage{
get_natural.predictors(x, variables = x, predictors = NULL)
}
\arguments{
\item{x}{a vector of character strings}

\item{variables}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Get natural predictors
}
\details{
if x is "a_1_f_1_aa.xoijj" returns c("a","1","f","1")
}
\examples{
TK<-TtableA
get_natural.predictors(x=sample(names(TtableA),5),variables=names(TtableA))
}
