% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{NAto0}
\alias{NAto0}
\title{Recoding of NAs to 0 or "NA"}
\usage{
NAto0(tableA)
}
\arguments{
\item{tableA}{a dataframe}
}
\value{
a dataframe
}
\description{
Recoding of NAs to 0 or "NA"
}
\details{
for synthetisation to run, missing values are treated as a special factor level for factor variables,
or as 0 for continuous variables. To avoid issues, for continuous variables, a missing indicator is also created.
}
\examples{
toto<-cars
toto$speed[sample(nrow(cars),3)]<-NA
NAto0(toto)
}
