% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceT_f.R
\name{reduceT_f}
\alias{reduceT_f}
\title{Reverse augmentT_f: Function that will convert cell and marginal ratios and overall total to cell values}
\usage{
reduceT_f(.data, variables, verbose = FALSE, hack = TRUE)
}
\arguments{
\item{.data}{data frame to "reduce"}

\item{variables}{list of variable names roots}

\item{verbose}{(default FALSE) if verbose, the formulae to compute the new variables is printed.}

\item{hack}{(default TRUE)}
}
\description{
Reverse augmentT_f: Function that will convert cell and marginal ratios and overall total to cell values
}
\details{
this functions looks for the Augmentation parameters in the package object Augmentparameters[[tablename]]$percent
For each variable listed in Augmentparameters[[tablename]]$percent, it looks for the corresponding variable in
.data and computes cell values from cell and marginal ratios and overall total
}
