% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCompare.R
\name{runCompare}
\alias{runCompare}
\title{Shiny App to visualize Data}
\usage{
runCompare(package1 = NULL, package2 = NULL)
}
\description{
Shiny App to visualize Data
}
\examples{
package1<-NULL
package2<-NULL
runCompare()
}
