% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{Sparameters.default.f}
\alias{Sparameters.default.f}
\title{Default synthetisation parameters based on variable names}
\usage{
Sparameters.default.f(
  ref.table,
  asis = NULL,
  notpredictor = NULL,
  variables = Sparameters.variables.reorder.default(names(ref.table)),
  predictors.matrix = predictor.matrix.default(variables)[!is.element(variables, asis),
    !is.element(variables, notpredictor)],
  splittingvar = NULL,
  moresplits = NULL,
  preferredmethod = "ctree",
  splithreshold = 100,
  defaultsynparameters = list(smoothing = "none", importance = TRUE, keep.forest = TRUE,
    minbucket = 30)
)
}
\arguments{
\item{ref.table}{a dataframe}

\item{asis}{a vector of character strings, indicating which variables to keep as is.}

\item{notpredictor}{a vector of character strings, indicating which variables are not supposed to be used as predictors.}

\item{variables}{a vector of character strings, indicating the variables to synthesize. Order is important.}

\item{predictors.matrix}{a predictor matrix. Number of rows is the number of variables to synthesize, number of columns is all the variables from ref.table}

\item{moresplits}{an object of class moresplist (not defined yet)}

\item{preferredmethod:}{"rf" for random forest or "ctree" for classification tree}

\item{defaultparameters}{a list indicating default parameters for synthpop synthesisation functions, for example ntree=5, smoothing="none"}
}
\value{

}
\description{
Default synthetisation parameters based on variable names
}
\details{
creates default synthetisation parameters
Some rules: parents variable are potential predictors of their children,
synthetisation is conditional to missingindicators,
synthetisation is conditional to presence in cell
}
\examples{
data(TtableA)
ATtableA<-augmentT_f(TtableA,variablespct="AA.cont1",variablesmax="AA.present")
ref.table<-ATtableA
Spa<-Sparameters.default.f(ref.table=ATtableA)
names(Spa)<-lapply(Spa,function(x){x$variable})
Spa$AA.present_La_Lb
Spa$AA.cont1_La_Lb
}
