% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2.R
\name{Sparameters.variables.reorder.default}
\alias{Sparameters.variables.reorder.default}
\title{General Default ordering of variables for synthetisation based on name of the variable.}
\usage{
Sparameters.variables.reorder.default(variables, orderwithinorigin = NULL,
  id = NULL)
}
\arguments{
\item{variables}{vector of character strings, indicating names of variables}

\item{orderwithinorigin}{a list, see example}
}
\value{
a list.
}
\description{
General Default ordering of variables for synthetisation based on name of the variable.
}
\examples{
TK<-Tsampledata(TRUE)
Sparameters.variables.reorder.default(names(TK$TtableA))
#Second example: no transposing variables
TK<-Tsampledata(FALSE)
orderwithinorigin=c("AA.factor1","AA.factor2")
Sparameters.variables.reorder.default(names(TK$TtableA),orderwithinorigin)
}
